Skip installing 1password-cli if we don't need it

+++ home/.chezmoihooks/install-password-manager.sh
@@ -6,6 +6,9 @@ PATH="$HOMEBREW_PREFIX/bin:$HOME/.local/bin:$PATH"
 if command -v op >/dev/null 2>&1; then
     exit
 fi
+if ! grep -E '^ *is_(home|siemens) *= *true' ~/.config/chezmoi/chezmoi.toml; then
+    exit
+fi
 
 tmpdir=$(mktemp -d -t install-package-manager.XXXXXX)
 trap 'rm -rf "$tmpdir"' EXIT INT TERM
